Fix Code

+++ back-end/src/controllers/auth.controller.js
@@ -8,6 +8,7 @@ const emailService = require('../services/email.service');
 const redis = require('../config/redis');
 const { encrypt, decrypt } = require('../utils/cryption');
 const analyzeUser = require('../utils/analyzeUser');
+const e = require('express');
 
 // Login
 const login = async (req, res) => {
@@ -32,7 +33,7 @@ const login = async (req, res) => {
         const userId = user._id;
 
         // Generate Token
-        const token = jwt.sign({ userId: user._id}, process.env.JWT_SECRET, { expiresIn: '9000h' });
+        const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET, { expiresIn: '9000h' });
 
         // Add token in Redis list
         await redis.client.rPush(userId.toString(), token);
@@ -48,16 +49,20 @@ const forgotPassword = async (req, res) => {
     try {
         const { email } = req.body;
 
-        const user = await User.findOne({ email });
+        const user = await User.findOne({
+        });
         if (user) {
-            await redis.client.del(email);
-            const code = Math.floor(100000 + Math.random() * 900000);
+            const code = String(Math.floor(100000 + Math.random() * 900000));
             const ttl = 300;
             emailService.resetPassword(email, code);
-            const emailencrypt = encrypt(email);
-            await redis.client.set(emailencrypt, code, 'EX', ttl);
+
             const token = crypto.randomBytes(35).toString('hex').toUpperCase();
-            await redis.client.set(token, emailencrypt, 'EX', ttl);
+            await redis.client.set(code, email, {
+                EX: ttl
+            });
+            await redis.client.set(token, code, {
+                EX: ttl
+            });
             res.status(200).json({ message: 'Successful send email', token: token });
         } else {
             return res.status(400).json({ message: 'Email does not exist' });
@@ -72,21 +77,26 @@ const confirmResetPassword = async (req, res) => {
     try {
         const { token, code, password } = req.body;
 
-        const emailInRedis = await redis.client.get(token);
-        if (!emailInRedis) {
+        const codeInRedis = await redis.client.get(token);
+        if (!codeInRedis) {
+            return res.status(400).json({ message: 'Invalid token reset' });
+        }
+        const emailInRedis = await redis.client.get(codeInRedis);
+        console.log(emailInRedis,
+            codeInRedis,
+            code,
+        );
+        if (!emailInRedis || codeInRedis !== String(code)) {
             return res.status(400).json({ message: 'Invalid token reset or code' });
         }
-        const emailDecrypt = decrypt(emailInRedis);
-        const codeInRedis = await redis.client.get(emailInRedis);
-
-        if (code === codeInRedis) {
-            await redis.client.del(emailInRedis);
+        if (String(code) === codeInRedis && emailInRedis) {
             await redis.client.del(token);
+            await redis.client.del(codeInRedis);
         } else {
             return res.status(400).json({ message: 'Invalid code' });
         }
 
-        const user = await User.findOne({ email: emailDecrypt });
+        const user = await User.findOne({ email: emailInRedis });
         if (user) {
             user.password = password;
             await user.save();
